# backend/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Timetable API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/config/terminology": {
      "get": {
        "tags": [
          "Config"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TerminologyEntry"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TerminologyEntry"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TerminologyEntry"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TerminologyEntry"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/config/features": {
      "put": {
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "boolean"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "boolean"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/config/time-structure": {
      "put": {
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeStructureDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeStructureDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TimeStructureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeStructureDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/constraints/{code}": {
      "put": {
        "tags": [
          "Config"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintSettingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintSettingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintSettingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/config/audit": {
      "get": {
        "tags": [
          "Config"
        ],
        "parameters": [
          {
            "name": "entityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfAuditEntryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/config/import": {
      "post": {
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfigRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfigRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfigChangePreview"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/apply-template": {
      "post": {
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyTemplateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyTemplateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfigChangePreview"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/save-as-template": {
      "post": {
        "tags": [
          "Config"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAsTemplateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAsTemplateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAsTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplyTemplateRequest": {
        "required": [
          "dryRun",
          "templateCode"
        ],
        "type": "object",
        "properties": {
          "templateCode": {
            "type": "string"
          },
          "dryRun": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AuditEntryDto": {
        "required": [
          "action",
          "at",
          "entityId",
          "entityType",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "action": {
            "$ref": "#/components/schemas/ChangeAction"
          },
          "beforeJson": {
            "type": "string",
            "nullable": true
          },
          "afterJson": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ChangeAction": {
        "enum": [
          "Created",
          "Updated",
          "Deleted"
        ],
        "type": "string"
      },
...
      "ConfigChangePreview": {
        "required": [
          "action",
          "area",
          "code"
        ],
        "type": "object",
        "properties": {
          "area": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "action": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ConstraintSettingRequest": {
        "required": [
          "severity",
          "weight"
        ],
        "type": "object",
        "properties": {
          "severity": {
            "$ref": "#/components/schemas/ConstraintSeverity"
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          },
          "parameters": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConstraintSeverity": {
        "enum": [
          "Off",
          "Soft",
          "Hard"
        ],
        "type": "string"
      },
...
      "DayOverrideDto": {
        "required": [
          "dayOfWeek",
          "disabled",
          "slotIndex"
        ],
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "integer",
            "format": "int32"
          },
          "slotIndex": {
            "type": "integer",
            "format": "int32"
          },
          "disabled": {
            "type": "boolean"
          },
          "start": {
            "type": "string",
            "nullable": true
          },
          "end": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ImportConfigRequest": {
        "required": [
          "bundle",
          "dryRun"
        ],
        "type": "object",
        "properties": {
          "bundle": { },
          "dryRun": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "PagedResultOfAuditEntryDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEntryDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PeriodDto": {
        "required": [
          "end",
          "index",
          "isBreak",
          "start"
        ],
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "nameAr": {
            "type": "string",
            "nullable": true
          },
          "nameEn": {
            "type": "string",
            "nullable": true
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "isBreak": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SaveAsTemplateRequest": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "nameAr": {
            "type": "string",
            "nullable": true
          },
          "nameEn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ShiftDto": {
        "required": [
          "code",
          "firstSlot",
          "id",
          "lastSlot"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "nameAr": {
            "type": "string",
            "nullable": true
          },
          "nameEn": {
            "type": "string",
            "nullable": true
          },
          "firstSlot": {
            "type": "integer",
            "format": "int32"
          },
          "lastSlot": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TerminologyEntry": {
        "required": [
          "key"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "ar": {
            "type": "string",
            "nullable": true
          },
          "en": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TimeStructureDto": {
        "required": [
          "dayOverrides",
          "periods",
          "shifts",
          "weekCycleLabels",
          "weekCycleLength",
          "weekStartDay",
          "workingDays"
        ],
        "type": "object",
        "properties": {
          "workingDays": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "weekStartDay": {
            "type": "integer",
            "format": "int32"
          },
          "weekCycleLength": {
            "type": "integer",
            "format": "int32"
          },
          "weekCycleLabels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodDto"
            }
          },
          "shifts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShiftDto"
            }
          },
          "dayOverrides": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOverrideDto"
            }
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}